+++ src/Linna/DataMapper/NullUuidlDomainObject.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Linna Framework.
+ *
+ * @author Sebastian Rapetti <sebastian.rapetti@alice.it>
+ * @copyright (c) 2018, Sebastian Rapetti
+ * @license http://opensource.org/licenses/MIT MIT License
+ */
+declare(strict_types=1);
+
+namespace Linna\DataMapper;
+
+/**
+ * NullUuidDomainObject.
+ */
+class NullUuidDomainObject extends UuidDomainObjectAbstract
+{
+}