+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 #### Data Mapper
+* `Linna\DataMapper\NullUuidDomainObject` interface
 * `Linna\DataMapper\UuidDomainObjectAbstract` class
 * `Linna\DataMapper\UuidDomainObjectInterface` interface
 * `Linna\DataMapper\UuidMapperAbstract` class
@@ -21,6 +22,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Router
 * `Linna\Router\Router->parseQueryStringOnRewriteModeOn` protected property
+
+### Fixed
+* `Linna\DataMapper\MapperAbstract->save()` update after insert
 
 ## [v0.26.0](https://github.com/linna/framework/compare/v0.25.0...v0.26.0) - 2019-08-05